+++ CHANGELOG.md
@@ -5,6 +5,49 @@ See [Conventional Commits](Https://conventionalcommits.org) for commit guideline
 
 <!-- changelog -->
 
+## [3.0.2](https://github.com/spandex-project/spandex/compare/3.0.1...3.0.2) (2020-07-13)
+
+
+
+
+### Bug Fixes:
+
+* add a name to failed traces (#116)
+
+## [3.0.1](https://github.com/spandex-project/spandex/compare/3.0.0...3.0.1) (2020-05-14)
+
+
+
+
+### Bug Fixes:
+
+* configure sender in tests
+
+## [3.0.0](https://github.com/spandex-project/spandex/compare/2.4.4...3.0.0) (2020-05-14)
+### Breaking Changes:
+
+* allow headers to be passed into `Spandex.distributed_context/2` (#113)
+
+
+
+## [2.4.4](https://github.com/spandex-project/spandex/compare/2.4.3...2.4.4) (2020-4-28)
+
+
+
+
+### Bug Fixes:
+
+* Set Logger.metadata when `continue_trace/3` is called as well (#111)
+
+## [2.4.3](https://github.com/spandex-project/spandex/compare/2.4.2...2.4.3) (2020-3-25)
+
+
+
+
+### Bug Fixes:
+
+* No unmatched returns (#109)
+
 ## [2.4.2](https://github.com/spandex-project/spandex/compare/2.4.1...2.4.2) (2020-1-13)